Db type configuration

+++ Telamingo.Domain/AggregateModels/CountryAggregate/EnvironmentalInformation.cs
@@ -5,8 +5,8 @@ namespace Telamingo.Domain.AggregateModels.CountryAggregate
     public class EnvironmentalInformation: Entity
     {
         public int CountryId { get; set; }
-        public double ReligiousIssues { get; set; }
-        public double Immigration { get; set; }
-        public double SocialRelationshipMatching { get; set; }
+        public int ReligiousIssues { get; set; }
+        public int Immigration { get; set; }
+        public int SocialRelationshipMatching { get; set; }
     }
 }
